PHP to calculate time
Here is a file that calculates the best time. Right now it only puts the best day data in an array where the keys are the weekdays, and the values are how many times the team members choose that weekday. For it to work, it needs to be stored in the includes folder. Also it doesn't link to any other pages so you have to manually put the address in your search bar to access it.